Created unit tests for get todo item by id method. Added additional exceptions in the logic.

+++ src/Logic/Logic.Core/TodoListLogic.cs
@@ -41,6 +41,10 @@
             {
                 throw new ArgumentException("Value can't be null or whitespace.", nameof(id));
             }
+            if (id.Length > 4)
+            {
+                throw new ArgumentException("Value can't have more that 4 characters.", nameof(id));
+            }
             var todo = await Repository.GetByIdAsync(id) ?? throw new EntityNotFoundException(id);
             return await Repository.DeleteTodoAsync(id);
         }
@@ -56,6 +60,9 @@
             {
                 throw new ArgumentException("Value can't be null or whitespace.", nameof(id));
             }
+            if (id.Length > 4) {
+                throw new ArgumentException("Value can't have more that 4 characters.", nameof(id));
+            }
             return await Repository.GetByIdAsync(id);
         }
         /// <inheritdoc />
@@ -64,6 +71,10 @@
             if (string.IsNullOrEmpty(id) || string.IsNullOrWhiteSpace(id))
             {
                 throw new ArgumentException("Value can't be null or whitespace.", nameof(id));
+            }
+            if (id.Length > 4)
+            {
+                throw new ArgumentException("Value can't have more that 4 characters.", nameof(id));
             }
             var todo = await GetTodoByIdAsync(id) ?? throw new EntityNotFoundException(id);
             if (todo.CompletedAt != null)
@@ -76,11 +87,15 @@
         /// <inheritdoc />
         public async Task<TodoItemModel?> UpdateTodoAsync(string id, TodoItemModel updateTodoItem)
         {
-            var todo = await Repository.GetByIdAsync(id) ?? throw new EntityNotFoundException(id);
             if (string.IsNullOrEmpty(id) || string.IsNullOrWhiteSpace(id))
             {
                 throw new ArgumentException("Value can't be null or whitespace.", nameof(id));
             }
+            if (id.Length > 4)
+            {
+                throw new ArgumentException("Value can't have more that 4 characters.", nameof(id));
+            }
+            var todo = await Repository.GetByIdAsync(id) ?? throw new EntityNotFoundException(id);
             if (string.IsNullOrEmpty(updateTodoItem.Title) || string.IsNullOrWhiteSpace(updateTodoItem.Title))
             {
                 throw new ArgumentException("Value can't be null or whitespace.", nameof(updateTodoItem.Id));
